CLIPSeg/requirements.txt
transformers
pillow
torch
matplotlib